[Rust] Extract test functions to the integration testing lib to share
